move api routes to router
